fix issue with generating static params
