+++ app/posts/[slug]/page.tsx
@@ -1,15 +1,19 @@
-import { fetchStoryBySlug, pageVersionProps } from "@/lib/api";
+import { fetchStoryBySlug, fetchAllPosts } from "@/lib/api";
 import { notFound } from 'next/navigation';
-import { StoryblokComponent, getStoryblokApi } from "@storyblok/react/rsc";
-
+import { StoryblokComponent } from "@storyblok/react/rsc";
 import Post from "@/components/templates/post/Post";
- 
+
+import { PostStoryPreviewProps } from "@/lib/types/posts"; 
+
 export default async function Page({
   params,
 }: {
   params: { slug: string };
 }) {
   const { data } = await fetchStoryBySlug(`posts/${params.slug}`);
+  if(data.error){
+    notFound();
+  }
   return !data.error ? 
     <Post title={ data.story.name } date={ data.story.published_at } editURL={`https://app.storyblok.com/#/me/spaces/${process.env.storyblokSpaceID}/stories/0/0/${data.story.id}`}>
       <StoryblokComponent blok={data.story.content} /></Post> : null;
@@ -20,36 +24,21 @@ export async function generateMetadata({
 }: {
   params: { slug: string };
 }) {
-  // notFound();
-  
-  
   const { data } = await fetchStoryBySlug(`posts/${params.slug}`);
   if(data.error){
     notFound();
   }
   return {
-    title: data.story.name
+    title: `${data.story.name}`
   }
 }
 
-/* 
-export async function generateStaticParams() {
-  const storyblokApi = getStoryblokApi();
-  let { data } = await storyblokApi.get("cdn/links/", {
-    version: process.env.storyblokPageVersion as pageVersionProps,
-    starts_with: 'posts/',
-  });
 
+export async function generateStaticParams() {
+  const data = await fetchAllPosts();
   let paths: { slug: string }[] = [];
- 
-  Object.keys(data.links).forEach((linkKey) => {
-    if (data.links[linkKey].is_folder) {
-      return;
-    }
-    const slug = data.links[linkKey].slug;
-   
-    paths.push({ slug: slug.replace('posts/', '') });
-  });
+  {data.stories.map((post: PostStoryPreviewProps) => { 
+    paths.push({slug: post.full_slug.replace('posts/', '')});
+  })}
   return paths;
 }
-*/
