+++ lib/api.ts
@@ -1,6 +1,7 @@
 import { getStoryblokApi } from "@storyblok/react/rsc";
 export type pageVersionProps = "published" | "draft";
 
+
 export async function fetchStoryBySlug(slug:string) {
   const storyblokApi = getStoryblokApi();
   try{
@@ -12,5 +13,12 @@ export async function fetchStoryBySlug(slug:string) {
   }
 }
 
-
-
+export async function fetchAllPosts(){
+  const storyblokApi = getStoryblokApi();
+  const { data } = await storyblokApi.get(`cdn/stories`, {
+    version: process.env.storyblokPageVersion as pageVersionProps,
+    starts_with: 'posts/',
+    is_startpage: false
+  });
+  return data;
+}
