+++ lib/types/posts.ts
@@ -1,0 +1,9 @@
+export interface PostStoryPreviewProps {
+  name: string;
+  uuid: string;
+  published_at: string;
+  full_slug: string;
+  content: {
+    description: string;
+  }
+}
